[advanced] 타입 분리
